+++ src/games/definitions.rs
@@ -83,4 +83,5 @@ pub static GAMES: Map<&'static str, Game> = phf_map! {
     "ut" => game!("Unreal Tournament", 7778, Protocol::Gamespy(GameSpyVersion::One)),
     "vr" => game!("V Rising", 27016, Protocol::Valve(SteamApp::VR)),
     "jc2mp" => game!("Just Cause 2: Multiplayer", 7777, Protocol::PROPRIETARY(ProprietaryProtocol::JC2MP)),
+    "warsow" => game!("Warsow", 44400, Protocol::Quake(QuakeVersion::Three)),
 };