+++ src/protocols/quake/two.rs
@@ -14,9 +14,10 @@ use super::QuakePlayerType;
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct Player {
-    pub frags: i16,
+    pub score: i32,
     pub ping: u16,
     pub name: String,
+    pub address: Option<String>,
 }
 
 impl QuakePlayerType for Player {
@@ -29,7 +30,8 @@ impl CommonPlayer for Player {
     fn as_original(&self) -> GenericPlayer { GenericPlayer::QuakeTwo(self) }
 
     fn name(&self) -> &str { &self.name }
-    // TODO: Maybe frags is score?
+
+    fn score(&self) -> Option<u32> { Some(self.score.try_into().unwrap_or(0)) }
 }
 
 pub(crate) struct QuakeTwo;
@@ -42,7 +44,7 @@ impl QuakeClient for QuakeTwo {
 
     fn parse_player_string(mut data: Iter<&str>) -> GDResult<Self::Player> {
         Ok(Player {
-            frags: match data.next() {
+            score: match data.next() {
                 None => Err(GDError::PacketBad)?,
                 Some(v) => v.parse().map_err(|_| GDError::PacketBad)?,
             },
@@ -53,6 +55,10 @@ impl QuakeClient for QuakeTwo {
             name: match data.next() {
                 None => Err(GDError::PacketBad)?,
                 Some(v) => remove_wrapping_quotes(v).to_string(),
+            },
+            address: match data.next() {
+                None => None,
+                Some(v) => Some(remove_wrapping_quotes(v).to_string()),
             },
         })
     }