+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Who knows what the future holds...
 ### Changes:
 Protocols:
 - GameSpy 2 support.
+- Quake 2: Added Optional address field to Player.
 
 Generic query:
 - Added generic queries (by [@Douile](https://github.com/Douile)) which come with a common struct for the response fields.
@@ -13,8 +14,12 @@ Generic query:
 Games:
 - [Halo: Combat Evolved](https://en.wikipedia.org/wiki/Halo:_Combat_Evolved) support.
 - [Just Cause 2: Multiplayer](https://store.steampowered.com/app/259080/Just_Cause_2_Multiplayer_Mod/) support.
+- [Warsow](https://warsow.net/) support.
 
 ### Breaking...
+Protocols:
+- Quake 2: Renamed the players frags field to score to be more inline with the other protocols.
+
 Crate:
 - Fixed crate's `rust-version`, it is now `1.60.0` (was `1.56.1`)